Use tcp_recv_bufsize to compute HttpStateData::inBuf capacity + polish
All other restrictions aside, the ideal zero-waste HttpStateData::inBuf
space size would be equal to the amount of payload data in the kernel
socket buffer. That ideal scenario is unachievable in practice because
we do not know, in general, how many TCP payload bytes the kernel socket
buffer has for us until we start reading. The next best space size would
be equal to the kernel socket buffer size -- we would never run out of
space while not allocating too much extra space either.

There are several options for determining the kernel socket space size.
Prior to this change, the branch code was using SQUID_TCP_SO_RCVBUF, but
that value has several serious problems:

    * cannot be changed without rebuilding Squid
    * determined on the build machine rather than on deployment host
    * cannot reflect deployment host configuration changes
    * ignores Squid's own tcp_recv_bufsize configuration

We are now using tcp_recv_bufsize instead of SQUID_TCP_SO_RCVBUF. I have
adjusted tcp_recv_bufsize documentation accordingly.

Alternatives to consider:

    1. A getsockopt(SO_RCVBUF) call.
    2. A new dedicated squid.conf directive.

Also simplified Client::calcBufferSpaceToReserve(space, wantSpace) API
to drop its unnecessary and confusing parameters. The method, now called
calcAccumulationAllowance(), should not care about other space limits
that the caller may need to obey -- they are irrelevant to the "how much
more data can we accumulate" question that this method is supposed to
answer. We need to solve many problems to answer that question. We
should not distract ourselves with irrelevant information.

Also converted a few APIs to use more appropriate integer types. This
change is essentially a bug fix: After the calcAccumulationAllowance()
simplification described above, the method needed to supply an upper
bound for the StoreEntry::bytesWanted() call. The only correct value for
that is size_t::max(), and bytesWanted() itself supported that value.
However, bytesWanted() calls MemObject::mostBytesWanted() that was using
int instead of size_t, resulting in negative maximums bugs.

Also partially addressed HttpStateData::maybeReadVirginBody() XXX that
was stalling test transactions.

Also polished HttpStateData::maybeMakeSpaceAvailable(), primarily to
clarify each limit/constraint the method has to obey/balance.

The code still has many problems that need to be addressed.

+++ src/MemObject.h
@@ -107,8 +107,7 @@ public:
 
     /// the maximum number of additional message body bytes the entry may buffer
     /// while still obeying the current read_ahead_gap configuration
-    /// \returns a positive number of body bytes or zero
-    int64_t readAheadAllowance() const;
+    uint64_t readAheadAllowance() const;
 
     void addClient(store_client *);
     /* XXX belongs in MemObject::swapout, once swaphdrsz is managed
@@ -120,7 +119,7 @@ public:
     void trimSwappable();
     void trimUnSwappable();
     bool isContiguous() const;
-    int mostBytesWanted(int max, bool ignoreDelayPools) const;
+    size_t mostBytesWanted(size_t max, bool ignoreDelayPools) const;
     void setNoDelay(bool const newValue);
 #if USE_DELAY_POOLS
     DelayId mostBytesAllowed() const;
